+++ buf.h
@@ -12,7 +12,7 @@ namespace ici
  * The following portion of this file exports to ici.h. --ici.h-start--
  */
 
-extern int      ici_growbuf(int);
+extern int      ici_growbuf(size_t);
 
 /*
  * Ensure that 'ici_buf' points to enough memory to hold index 'n' (plus
@@ -23,7 +23,7 @@ extern int      ici_growbuf(int);
  *
  * This --func-- forms part of the --ici-api--.
  */
-inline int ici_chkbuf(int n) { return ici_bufz > n ? 0 : ici_growbuf(n); }
+inline int ici_chkbuf(size_t n) { return ici_bufz > n ? 0 : ici_growbuf(n); }
 
 /*
  * End of ici.h export. --ici.h-end--
